feat(MCTS): Apply changes of separated files
