+++ Extensions/RosettaTorch/Includes/MCTS/Selection/Selection.hpp
@@ -11,8 +11,8 @@
 #define ROSETTASTONE_TORCH_MCTS_SELECTION_HPP
 
 #include <MCTS/Policies/CreditPolicy.hpp>
+#include <MCTS/Policies/Selection/UCBPolicy.hpp>
 #include <MCTS/Policies/StageController.hpp>
-#include <MCTS/Policies/UCBPolicy.hpp>
 #include <MCTS/Selection/TraversedNodesInfo.hpp>
 #include <MCTS/Selection/TreeNode.hpp>
 #include <MCTS/Types.hpp>
